refactor: вынес старый класс дао в отдельный пакет
