+++ hw-5/src/main/java/ru/dankoy/hw5/core/dao/book/mergemanytomanybycode/BookDao.java
@@ -1,4 +1,4 @@
-package ru.dankoy.hw5.core.dao.book;
+package ru.dankoy.hw5.core.dao.book.mergemanytomanybycode;
 
 import java.util.List;
 import ru.dankoy.hw5.core.domain.Book;
