+++ hw-5/src/test/java/ru/dankoy/hw5/core/dao/book/BookDaoJdbcTest.java
@@ -15,6 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.context.annotation.Import;
 import ru.dankoy.hw5.core.dao.author.AuthorDaoJdbc;
+import ru.dankoy.hw5.core.dao.book.mergemanytomanybycode.BookDaoJdbc;
 import ru.dankoy.hw5.core.dao.genre.GenreDaoJdbc;
 import ru.dankoy.hw5.core.domain.Author;
 import ru.dankoy.hw5.core.domain.Book;
