+++ hw-5/src/test/java/ru/dankoy/hw5/core/service/book/BookServiceJdbcTest.java
@@ -15,7 +15,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.context.annotation.Import;
-import ru.dankoy.hw5.core.dao.book.BookDao;
+import ru.dankoy.hw5.core.dao.book.mergemanytomanybycode.BookDao;
 import ru.dankoy.hw5.core.domain.Author;
 import ru.dankoy.hw5.core.domain.Book;
 import ru.dankoy.hw5.core.domain.Genre;
